chore: make pipeline builder more object [part-2]
